+++ addons/account/static/src/xml/legacy_account_payment.xml
@@ -2,7 +2,7 @@
 
 <templates xml:space="preserve">
 
-    <t t-name="AccountPaymentInfo" owl="1">
+    <t t-name="ShowPaymentInfo">
         <div>
             <t t-if="outstanding">
                 <div>
@@ -10,7 +10,7 @@
                 </div>
             </t>
             <table style="width:100%;">
-                <t t-foreach="lines" t-as="line" t-key="line_index">
+                <t t-foreach="lines" t-as="line">
                     <tr>
                     <t t-if="outstanding">
                         <td>
@@ -20,8 +20,7 @@
                                t-att-data-id="line.id"
                                style="margin-right: 0px; padding-left: 5px; padding-right: 5px;"
                                href="#"
-                               data-bs-toggle="tooltip"
-                               t-on-click.prevent="() => this.outstandingCreditAssign(line.id)">Add</a>
+                               data-bs-toggle="tooltip">Add</a>
                         </td>
                         <td style="max-width: 11em;">
                             <a t-att-title="line.date"
@@ -36,7 +35,7 @@
                     </t>
                     <t t-if="!outstanding">
                         <td>
-                           <a role="button" tabindex="0" class="js_payment_info fa fa-info-circle" t-att-index="line_index" style="margin-right:5px;" aria-label="Info" title="Journal Entry Info" data-bs-toggle="tooltip" t-on-click.stop="(ev) => this.onInfoClick(ev, line_index)"></a>
+                           <a role="button" tabindex="0" class="js_payment_info fa fa-info-circle" t-att-index="line.index" style="margin-right:5px;" aria-label="Info" title="Journal Entry Info" data-bs-toggle="tooltip"></a>
                         </td>
                         <td t-if="!line.is_exchange">
                             <i class="o_field_widget text-start o_payment_label">Paid on <t t-esc="line.date"></t></i>
@@ -73,43 +72,38 @@
         </div>
     </t>
 
-    <t t-name="AccountPaymentPopOver" owl="1">
-        <div class="account_payment_popover">
-            <h3 t-if="props.title" class="o_popover_header"><t t-esc="props.title"/></h3>
-            <div class="px-2">
-                <div>
-                    <table>
-                        <tr>
-                            <td><strong>Amount: </strong></td>
-                            <td>
-                                <t t-esc="props.amount_company_currency"></t>
-                                <t t-if="props.amount_foreign_currency">
-                                    (<span class="fa fa-money"/> <t t-esc="props.amount_foreign_currency"/>)
-                                </t>
-                            </td>
-                        </tr>
-                        <tr>
-                            <td><strong>Memo: </strong></td>
-                            <td>
-                                <div style="width: 200px; word-wrap: break-word">
-                                    <t t-esc="props.ref"/>
-                                </div>
-                            </td>
-                        </tr>
-                        <tr>
-                            <td><strong>Date: </strong></td>
-                            <td><t t-esc="props.date"/></td>
-                        </tr>
-                        <tr>
-                            <td><strong>Journal: </strong></td>
-                            <td><t t-esc="props.journal_name"/><span t-if="props.payment_method_name"> (<t t-esc="props.payment_method_name"/>)</span></td>
-                        </tr>
-                    </table>
-                </div>
-                <button class="btn btn-sm btn-primary js_unreconcile_payment float-start" style="margin-top:5px; margin-bottom:5px;" groups="account.group_account_invoice" t-on-click="() => props._onRemoveMoveReconcile(props.move_id, props.partial_id)">Unreconcile</button>
-                <button class="btn btn-sm btn-secondary js_open_payment float-end" style="margin-top:5px; margin-bottom:5px;" t-on-click="() => props._onOpenPaymentOrMove(props.account_payment_id ? 'account.payment' : 'account.move', props.account_payment_id or props.move_id)">View</button>
-            </div>
+    <t t-name="PaymentPopOver">
+        <div>
+            <table>
+                <tr>
+                    <td><strong>Amount: </strong></td>
+                    <td>
+                        <t t-esc="amount_company_currency"></t>
+                        <t t-if="amount_foreign_currency">
+                            (<span class="fa fa-money"/> <t t-esc="amount_foreign_currency"/>)
+                        </t>
+                    </td>
+                </tr>
+                 <tr>
+                    <td><strong>Memo: </strong></td>
+                    <td>
+                        <div style="width: 200px; word-wrap: break-word">
+                            <t t-esc="ref"/>
+                        </div>
+                    </td>
+                </tr>
+                <tr>
+                    <td><strong>Date: </strong></td>
+                    <td><t t-esc="date"/></td>
+                </tr>
+                <tr>
+                    <td><strong>Journal: </strong></td>
+                    <td><t t-esc="journal_name"/><span t-if="payment_method_name"> (<t t-esc="payment_method_name"/>)</span></td>
+                </tr>
+            </table>
         </div>
+        <button class="btn btn-sm btn-primary js_unreconcile_payment float-start" t-att-partial-id="partial_id" t-att-payment-id="payment_id" t-att-move-id="move_id" style="margin-top:5px; margin-bottom:5px;" groups="account.group_account_invoice">Unreconcile</button>
+        <button class="btn btn-sm btn-secondary js_open_payment float-end" t-att-payment-id="account_payment_id" t-att-move-id="move_id" style="margin-top:5px; margin-bottom:5px;">View</button>
     </t>
 
 </templates>
